+++ amo2kinto/kinto.py
@@ -2,25 +2,6 @@ from functools import partial
 from kinto_http.exceptions import KintoException
 
 
-def update_schema_if_mandatory(response, config, patch_collection):
-    if 'details' in response:
-        collection_schema = {
-            'schema': response['details']['existing'].get('schema'),
-            'uiSchema': response['details']['existing'].get('uiSchema'),
-            'displayFields': response['details']['existing'].get(
-                'displayFields'),
-        }
-    else:
-        collection_schema = {
-            'schema': response['data'].get('schema'),
-            'uiSchema': response['data'].get('uiSchema'),
-            'displayFields': response['data'].get('displayFields'),
-        }
-
-    if config and collection_schema != config:
-        patch_collection(data=config)
-
-
 def get_kinto_records(kinto_client, bucket, collection, permissions,
                       config=None):
     """Return all the kinto records for this bucket/collection."""
@@ -39,11 +20,5 @@ def get_kinto_records(kinto_client, bucket, collection, permissions,
         if hasattr(e, 'response') and e.response.status_code == 403:
             # The user cannot create collection on this bucket, ignore the creation.
             pass
-        response = kinto_client.get_collection(id=collection, bucket=bucket)
-
-    patch_collection = partial(kinto_client.patch_collection,
-                               bucket=bucket, id=collection)
-
-    update_schema_if_mandatory(response, config, patch_collection)
 
     return kinto_client.get_records(bucket=bucket, collection=collection)